requirements.txt
Django==4.2
djangorestframework==3.15
bcrypt==4.0.1
PyJWT==2.6.0
psycopg2-binary==2.9.5
django-cors-headers==4.0.0  # если нужен CORS
